backend/payments/requirements.txt
fastapi
uvicorn[standard]
stripe
python-dotenv
pydantic
